chore: refactor lib
