+++ tools/ice-scripts/lib/commands/init/initProject.js
@@ -7,8 +7,8 @@ const uuid = require('uuid/v1');
 const rimraf = require('rimraf');
 const { checkAliInternal } = require('ice-npm-utils');
 
-const log = require('../utils/log');
-const download = require('./download');
+const log = require('../../utils/log');
+const download = require('../../utils/download');
 
 module.exports = ({ template, projectDir }) => {
   return download({
